+++ src/library/DIPS.Mobile.UI/Resources/Styles/Shared/TextStyle.cs
@@ -1,0 +1,21 @@
+namespace DIPS.Mobile.UI.Resources.Styles.Shared;
+
+public enum TextStyle
+{
+    None=0,
+    Body400=1,
+    Body300=2,
+    Body200=3,
+    Body100=4,
+    UI400=5,
+    UI300=6,
+    UI200=7,
+    UI100=8,
+    Header1000=9,
+    Header900=10,
+    Header800=11,
+    Header700=12,
+    Header600=13,
+    Header500=14,
+    SectionHeader=15
+}